requirements.txt
gspread==6.1.2
google-auth==2.29.0
requests==2.31.0
python-dateutil==2.9.0
anthropic==0.25.0
